Compare lowercase email when updating from ldap
Scenario:
    Nextcloud has an email with capital letters for a user. No idea why, as it seems we are storing it in lowercases.
    LDAP has the same email address.
    The user log in, calling processAttributes, which call updateEmail, which compare emails as they are stored ($currentEmail !== $email)
    But when we are about to update the email, we once again check old and new values, but this time we lower case the new value

This means that Nextcloud will try to update the email again and again because we compare an address email with its lower case equivalent. Remember that Nextcloud has an email with capital letter in its database.
So the user will receive a continuous stream of email signaling an email change has we later compare email as they are before triggering the event.

Signed-off-by: Louis Chemineau <louis@chmn.me>